seq_ordem_linha adicionando de 1000 em 1000

+++ src/model/db/db_orm.py
@@ -1,4 +1,5 @@
-from sqlalchemy import Column
+from decimal import Decimal
+from sqlalchemy import Column, Numeric
 from sqlalchemy import ForeignKey
 from sqlalchemy import Integer
 from sqlalchemy import String
@@ -42,7 +43,7 @@ class Lancamentos(Base):
 
     id = Column(Integer, primary_key=True, autoincrement="auto")
     # Para lançamentos na mesma data é possivel ordenar a posição deles baseado neste numero deve ser (1,2,3,4...)
-    seq_ordem_linha = Column(Integer)
+    seq_ordem_linha = Column(Numeric(10,5))
     nr_referencia = Column(String)
     descricao = Column(String)
     descricao_user = Column(String)
